pyproject.toml
[project]
name = "ecommerce-visual-pro"
version = "0.1.0"
description = "AI-powered e-commerce product photo optimization"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "celery>=5.3.6",
    "redis>=5.0.1",
    "pydantic-settings>=2.1.0",
    "httpx>=0.26.0",
    "pillow>=10.2.0",
    "python-multipart>=0.0.9",
    "google-cloud-storage>=2.14.0",
    "torch>=2.2.0",
    "transformers>=4.39.1,<4.49.0",
    "torchvision>=0.17.0",
    "scikit-image>=0.22.0",
    "numpy>=1.26.0",
]

[tool.uv]
dev-dependencies = [
    "ruff>=0.1.14",
    "mypy>=1.8.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
disable_error_code = ["untyped-decorator"]
